requirements.txt
aenum==3.1.11
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
apprise==1.2.1
APScheduler==3.9.1.post1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
Babel==2.11.0
Backtesting==0.3.3
beautifulsoup4==4.11.1
bidict==0.22.1
binance-connector==2.0.0
black==22.12.0
blosc==1.11.1
bokeh==3.0.3
cachetools==5.2.1
ccxt==2.6.47
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cheroot==9.0.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.1.6
docopt==0.6.2
fastapi==0.89.1
filelock==3.9.0
finta==1.3
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
frozendict==2.3.4
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
jaraco.functools==3.5.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
more-itertools==9.0.0
mpl-finance==0.10.1
mplfinance==0.12.9b7
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==0.4.3
numexpr==2.8.4
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandas-ta==0.3.14b0
pathspec==0.10.3
patsy==0.5.3
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
pluginlib==0.9.0
portalocker==2.7.0
prettytable==3.6.0
progressbar2==4.2.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
py-find-1st==1.1.5
py3cw==0.0.39
pyaes==1.6.1
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
pyserial-asyncio==0.6
python-binance==1.0.16
python-dateutil==2.8.2
python-engineio==4.3.4
python-rapidjson==1.6
python-socketio==5.7.2
python-telegram-bot==20.0
python-utils==3.4.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
questionary==1.10.0
rd==1.3.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.10.0
sdnotify==0.3.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
starlette==0.23.1
statsmodels==0.13.5
ta==0.10.2
TA-Lib==0.4.25
tables==3.7.0
tabulate==0.9.0
technical==1.3.0
Telethon==1.26.1
tenacity==8.1.0
threadpoolctl==3.1.0
three-commas==0.2.9
tomli==2.0.1
tornado==6.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
unicorn-binance-rest-api==1.5.0
unicorn-binance-websocket-api==1.41.0
unicorn-fy==0.12.2
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
xyzservices==2022.9.0
yarg==0.1.9
yarl==1.8.2
yfinance==0.2.4
zope.interface==5.5.2
